possible constructor from a while back
it is just the annotation and javadoc (which is rather verbose, and
why I am bothering to stash it).

constructor based injection is nice because you can put more logic
into the plain old java code, handles final fields, plays nicely
with IDEs (warnings etc), and is less dependent on the injection
framework (which if nothing else, can be sometimes useful for
tests).

however, there are complications -- mostly solvable, but involved:
- constructor parameter names are non-trivial to access at runtime
- fields that still need to be injected are non-trivial to
initialize with the same timing/ ordering of normal non-static
initializers. ie. injected after super-class constructors run and
before that class's constructor runs.
- should almost certainly be using javax inject
- unclear semantics surrounding when it should be used, whether
the parameters are required or can be null, non-config-codec usage