simple support for field renaming
there is no remapping or interacting with _primary et al. but it
is inherited along with other alias options, and there is already
the base line hocon built in substitutions/ includes so I feel it
is likely enough that a need for more complex features would
strongly indicate too much abstraction from the code/ base case.