error reporting with line numbers
the resolution is a little limited in some cases. eg. there is
a bug in jackson that prevents splitting up arrays in some cases.
it is fixed in master so hopefully they will release soon.

also includes a number of extra utilities and a general purpose
hocon-structure (not text) parser.