remove 'autocollection' option from FieldConfig
jackson supports that functionality but at a global setting level
only. Since I was considering turning it on by default anyway and
haven't yet run into a case where it was undesirable, I have just
enabled that flag by default. If a case comes up where we need to
disable it, there are plenty of reasonable options to explore.