Implement @FieldConfig(narrow = true) annotation.
This patch is a little unorthodox but it allows for the
storage of some primitive types in a more narrow type for
serialization and deserialization. This may result in the loss
of information but the intent is that is used only to preserve
backwards compatibility. The user of this annotation is required
to ensure that large values are not used during narrowing.

The intent is to apply this annotation
here: https://github.com/addthis/hydra/blob/spiegel/long-node-ids-cleanup/hydra-data/src/main/java/com/addthis/hydra/data/tree/AbstractTreeNode.java#L41
to allow for legacy pages that uses CodecBin2 encoding. New jobs
use the BytesCodable interface and allow for 64-bit node identifiers.
Any existing jobs will use 32-bit node identifiers. Ref T51939.